added saga folder and added getGifs.saga
